pyproject.toml
[project]
name = "csit-hack-backend"
version = "0.1.0"
description = "Backend for CSIT hackathon"
authors = [
    { name = "Artemiy Granat", email = "artemiigranat@gmail.com" },
    # { name = "Daniil Koveshnikov", email = "..." },
    # { name = "Vadim Posokhov", email = "..." },
    # { name = "Viktoriia Romanova", email = "..." }
]
dependencies = [
    "fastapi>=0.109.0",
    "pydantic>=2.5.3",
    "uvicorn>=0.25.0",
    "httpx>=0.26.0",
    "asgi_correlation_id>=4.3.0",
    "databases>=0.8.0",
    "asyncpg>=0.28.0",
    "bcrypt>=4.1.2",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.1.13",
    "pre-commit>=3.5.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages=["backend/src"]

[tool.ruff]
line-length = 79
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["I", "E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
